+++ Near/app/src/main/java/com/alarmy/near/data/di/DataStoreModule.kt
@@ -1,0 +1,25 @@
+package com.alarmy.near.data.di
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+// DataStore 확장 프로퍼티
+private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "auth_preferences")
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DataStoreModule {
+    @Provides
+    @Singleton
+    fun provideDataStore(
+        @ApplicationContext context: Context,
+    ): DataStore<Preferences> = context.dataStore
+}
